move TableId out of hbase package.
